rootcause_internals

Struct RawReport

pub struct RawReport { /* private fields */ }

A pointer to a [ReportData] that is guaranteed to point to an initialized instance of a [ReportData<C>] for some specific C, though we do not know which actual C it is.

However, the pointer is allowed to transition into a non-initialized state inside the RawReport::drop method.

The pointer is guaranteed to have been created using triomphe::Arc::into_raw.

We cannot use a triomphe::OffsetArc<ReportData<C>> directly, because that does not allow us to type-erase the C.

Implementations

impl RawReport

pub unsafe fn into_parts<C: 'static>( self, ) -> (C, Vec<RawReport>, Vec<RawAttachment>)

Deconstructs this report into its context, children, and attachments.

Safety

The caller must ensure:

The type C matches the actual context type stored in the [ReportData]

This is the only existing reference pointing to the inner [ReportData]

impl RawReport

pub fn new<C, H>( context: C, children: Vec<RawReport>, attachments: Vec<RawAttachment>, ) -> Self
where C: 'static, H: ContextHandler<C>,

Creates a new RawReport with the specified handler, context, children, and attachments.

The created report will have the supplied context type and handler type. It will also have a strong count of 1.

pub fn as_ref(&self) -> RawReportRef<'_>

Returns a reference to the [ReportData] instance.

pub unsafe fn as_mut(&mut self) -> RawReportMut<'_>

Returns a mutable reference to the [ReportData] instance.

Safety

The caller must ensure:

This is the only existing reference pointing to the inner [ReportData]. Specifically the strong count of the inner triomphe::Arc must be 1.